Move tree-people-base-vm into component
